lambda_code/requirements.txt
psycopg2-binary
PyPDF2